settings: split audio and video page
