+++ iina/SettingsAudioLocalizable.strings
@@ -1,0 +1,55 @@
+/* Audio threads (reuses video threads label/desc text) */
+"videoThreads.label" = "Number of threads";
+"videoThreads.desc" = "Default: 0 (Auto)";
+
+/* Audio driver */
+"audioDriverEnableAVFoundation.label" = "Audio driver";
+"audioDriverEnableAVFoundation.items.0" = "Core Audio";
+"audioDriverEnableAVFoundation.items.0.desc" = "The default audio driver.";
+"audioDriverEnableAVFoundation.items.1" = "AVFoundation";
+"audioDriverEnableAVFoundation.items.1.desc" = "AVFoundation is a new, experimental audio driver offering Spatial Audio support. However, you may encounter some issues when using AVFoundation, such as audio delays.";
+
+/* Volume */
+"enableInitialVolume.label" = "Initial volume";
+"maxVolume.label" = "Maximum volume";
+"maxVolume.desc" = "(100-1000)";
+
+/* Gapless audio */
+"gaplessAudio.label" = "Gapless audio";
+"gaplessAudio.desc" = "Gapless audio";
+"gaplessAudio.items.0" = "Disabled";
+"gaplessAudio.items.1" = "Weak";
+"gaplessAudio.items.2" = "Strong";
+"gaplessAudio.items.0.desc" = "Playback of consecutive audio files may result in silence or disruption at the point of file change.";
+"gaplessAudio.items.1.desc" = "If the audio format the decoder output changes, the audio device is closed and reopened, disrupting gapless playback.";
+"gaplessAudio.items.2.desc" = "The audio device is opened using parameters chosen for the first file played and is then kept open for gapless playback. Following files may be resampled.";
+
+/* Audio device */
+"preferredAudioDevice.label" = "Preferred audio device";
+
+/* S/PDIF */
+"spdifAC3.label" = "AC3";
+"spdifDTS.label" = "DTS";
+"spdifDTSHD.label" = "DTS-HD";
+
+/* Replay gain */
+"replayGain.label" = "Gain adjustment";
+"replayGain.desc" = "Adjust volume gain according to ReplayGain values stored in the file metadata.";
+"replayGain.items.0" = "No adjustment";
+"replayGain.items.1" = "Track gain";
+"replayGain.items.2" = "Album gain";
+"replayGainPreamp.label" = "Preamp gain";
+"replayGainPreamp.desc" = "Pre-amplification gain to apply to the ReplayGain gain.";
+"replayGainClip.label" = "Allow the volume gain to clip";
+"replayGainClip.desc" = "If this option is not enabled, clipping will be prevented by lowering the gain.";
+"replayGainFallback.label" = "Fallback";
+"replayGainFallback.desc" = "Gain to apply if the file has no ReplayGain tags.";
+
+/* Section titles and inline labels */
+"$Hardware" = "Hardware";
+"$Volume" = "Volume";
+"$Other" = "Other";
+"$SPDIFOutput" = "S/PDIF output";
+"$PreferredLanguage" = "Preferred language";
+"$ReplayGain" = "Replay Gain";
+"$dB" = "dB";
