+++ iina/SettingsLocalizationKeysAudio.swift
@@ -1,0 +1,30 @@
+extension SettingsLocalization.Key {
+  static let audioDriverEnableAVFoundationLabel = SettingsLocalization.Key("audioDriverEnableAVFoundation.label")
+  static let spdifAC3Label = SettingsLocalization.Key("spdifAC3.label")
+  static let spdifDTSLabel = SettingsLocalization.Key("spdifDTS.label")
+  static let spdifDTSHDLabel = SettingsLocalization.Key("spdifDTSHD.label")
+  static let enableInitialVolumeLabel = SettingsLocalization.Key("enableInitialVolume.label")
+  static let gaplessAudioLabel = SettingsLocalization.Key("gaplessAudio.label")
+  static let replayGainLabel = SettingsLocalization.Key("replayGain.label")
+  static let replayGainItem0 = SettingsLocalization.Key("replayGain.items.0")
+  static let replayGainItem1 = SettingsLocalization.Key("replayGain.items.1")
+  static let replayGainItem2 = SettingsLocalization.Key("replayGain.items.2")
+  static let replayGainClipLabel = SettingsLocalization.Key("replayGainClip.label")
+  static let replayGainClipDesc = SettingsLocalization.Key("replayGainClip.desc")
+  static let preferredAudioDeviceLabel = SettingsLocalization.Key("preferredAudioDevice.label")
+  static let text_MaximumVolume = SettingsLocalization.Key("$MaximumVolume")
+  static let text_Audio = SettingsLocalization.Key("$Audio")
+  static let text_SPDIFOutput = SettingsLocalization.Key("$SPDIFOutput")
+  static let text_PreferredLanguage = SettingsLocalization.Key("$PreferredLanguage")
+  static let text_ReplayGain = SettingsLocalization.Key("$ReplayGain")
+  static let text_GainAdjustment = SettingsLocalization.Key("$GainAdjustment")
+  static let text_AdjustVolumeGainAccordingTo = SettingsLocalization.Key("$AdjustVolumeGainAccordingTo")
+  static let text_PreampGain = SettingsLocalization.Key("$PreampGain")
+  static let text_dB = SettingsLocalization.Key("$dB")
+  static let text_PreamplificationGainToApplyTo = SettingsLocalization.Key("$PreamplificationGainToApplyTo")
+  static let text_Fallback = SettingsLocalization.Key("$Fallback")
+  static let text_GainToApplyIfThe = SettingsLocalization.Key("$GainToApplyIfThe")
+  static let text_Hardware = SettingsLocalization.Key("$Hardware")
+  static let text_AudioOther = SettingsLocalization.Key("$Other")
+  static let text_Volume = SettingsLocalization.Key("$Volume")
+}
