+++ iina/SettingsLocalizationKeysVideo.swift
@@ -22,37 +22,15 @@ extension SettingsLocalization.Key {
   static let toneMappingAlgorithmItem5 = SettingsLocalization.Key("toneMappingAlgorithm.items.5")
   static let toneMappingAlgorithmItem6 = SettingsLocalization.Key("toneMappingAlgorithm.items.6")
   static let toneMappingAlgorithmItem7 = SettingsLocalization.Key("toneMappingAlgorithm.items.7")
-  static let audioDriverEnableAVFoundationLabel = SettingsLocalization.Key("audioDriverEnableAVFoundation.label")
-  static let spdifAC3Label = SettingsLocalization.Key("spdifAC3.label")
-  static let spdifDTSLabel = SettingsLocalization.Key("spdifDTS.label")
-  static let spdifDTSHDLabel = SettingsLocalization.Key("spdifDTSHD.label")
-  static let enableInitialVolumeLabel = SettingsLocalization.Key("enableInitialVolume.label")
-  static let gaplessAudioLabel = SettingsLocalization.Key("gaplessAudio.label")
-  static let replayGainLabel = SettingsLocalization.Key("replayGain.label")
-  static let replayGainItem0 = SettingsLocalization.Key("replayGain.items.0")
-  static let replayGainItem1 = SettingsLocalization.Key("replayGain.items.1")
-  static let replayGainItem2 = SettingsLocalization.Key("replayGain.items.2")
-  static let replayGainClipLabel = SettingsLocalization.Key("replayGainClip.label")
-  static let replayGainClipDesc = SettingsLocalization.Key("replayGainClip.desc")
-  static let text_DefaultAuto = SettingsLocalization.Key("$DefaultAuto")
-  static let text_MaximumVolume = SettingsLocalization.Key("$MaximumVolume")
+  static let text_Decoding = SettingsLocalization.Key("$Decoding")
+  static let text_VideoColor = SettingsLocalization.Key("$VideoColor")
+  // Also used by the audio page for the audio threads input
   static let videoThreadsLabel = SettingsLocalization.Key("videoThreads.label")
   static let videoThreadsDesc = SettingsLocalization.Key("videoThreads.desc")
   static let text_Video = SettingsLocalization.Key("$Video")
-  static let preferredAudioDeviceLabel = SettingsLocalization.Key("preferredAudioDevice.label")
-  static let text_Audio = SettingsLocalization.Key("$Audio")
-  static let text_SPDIFOutput = SettingsLocalization.Key("$SPDIFOutput")
-  static let text_PreferredLanguage = SettingsLocalization.Key("$PreferredLanguage")
-  static let text_EnableToneMappingImagesOnto = SettingsLocalization.Key("$EnableToneMappingImagesOnto")
   static let text_nits = SettingsLocalization.Key("$nits")
+  static let text_DefaultAuto = SettingsLocalization.Key("$DefaultAuto")
+  static let text_EnableToneMappingImagesOnto = SettingsLocalization.Key("$EnableToneMappingImagesOnto")
   static let text_Algorithm = SettingsLocalization.Key("$Algorithm")
   static let text_AlgorithmToUseForTone = SettingsLocalization.Key("$AlgorithmToUseForTone")
-  static let text_ReplayGain = SettingsLocalization.Key("$ReplayGain")
-  static let text_GainAdjustment = SettingsLocalization.Key("$GainAdjustment")
-  static let text_AdjustVolumeGainAccordingTo = SettingsLocalization.Key("$AdjustVolumeGainAccordingTo")
-  static let text_PreampGain = SettingsLocalization.Key("$PreampGain")
-  static let text_dB = SettingsLocalization.Key("$dB")
-  static let text_PreamplificationGainToApplyTo = SettingsLocalization.Key("$PreamplificationGainToApplyTo")
-  static let text_Fallback = SettingsLocalization.Key("$Fallback")
-  static let text_GainToApplyIfThe = SettingsLocalization.Key("$GainToApplyIfThe")
 }
