--- iina/SettingsVideoAudioLocalizable.strings
@@ -1,134 +1,0 @@
-/* id="" */
-"videoThreads.desc" = "Default: 0 (Auto)";
-/* id="rFu-rL-EJC" */
-"hardwareDecoder.items.0" = "Disabled";
-/* id="tAQ-oU-1AJ" */
-"hardwareDecoder.items.1" = "Auto";
-/* id="0DK-uZ-duj" */
-"hardwareDecoder.items.2" = "Auto (Copy)";
-/* id="YUr-aJ-xfC" */
-"forceDedicatedGPU.label" = "Force dedicated GPU";
-/* id="gCv-sJ-uQJ" */
-"forceDedicatedGPU.desc" = "Always use the dedicated GPU for rendering (if it exists). This can improve performance but may reduce battery life.";
-/* id="3g4-jW-uJd" */
-"loadIccProfile.label" = "Load ICC profile";
-/* id="M2M-of-gjd" */
-"loadIccProfile.desc" = "Load the ICC profile for current display and use it to transform video RGB to screen output. (Only for SDR mode)";
-/* id="R7Z-9c-vPJ" */
-"enableHdrSupport.label" = "Enable HDR mode";
-/* id="ZgQ-bc-yEE" */
-"enableHdrSupport.desc" = "Enable HDR mode by default when playing HDR videos.";
-/* id="a5Y-gU-JTj" */
-"enableToneMapping.label" = "Enable tone mapping";
-/* id="" */
-"toneMappingAlgorithm.label" = "Algorithm";
-/* id="uey-Ao-wwZ" */
-"toneMappingAlgorithm.items.0" = "auto";
-/* id="hpU-bf-sNW" */
-"toneMappingAlgorithm.items.1" = "clip";
-/* id="8QO-fX-FbS" */
-"toneMappingAlgorithm.items.2" = "mobius";
-/* id="1Kp-GN-lWF" */
-"toneMappingAlgorithm.items.3" = "reinhard";
-/* id="n6V-Hw-yNt" */
-"toneMappingAlgorithm.items.4" = "hable";
-/* id="Q01-SJ-M3x" */
-"toneMappingAlgorithm.items.5" = "bt.2390";
-/* id="6Iq-cU-A17" */
-"toneMappingAlgorithm.items.6" = "gamma";
-/* id="yd0-W0-bHk" */
-"toneMappingAlgorithm.items.7" = "linear";
-/* id="1uU-MC-7Yh" */
-"spdifAC3.label" = "AC3";
-/* id="Le8-wj-Pei" */
-"spdifDTS.label" = "DTS";
-/* id="I34-op-NZK" */
-"spdifDTSHD.label" = "DTS-HD";
-/* id="g4c-vG-URh" */
-"enableInitialVolume.label" = "Initial volume";
-/* id="" */
-"replayGain.label" = "Gain adjustment";
-/* id="k7b-6o-lCK" */
-"replayGain.items.0" = "No adjustment";
-/* id="w6E-lw-g7g" */
-"replayGain.items.1" = "Track gain";
-/* id="ZJJ-1o-ZGD" */
-"replayGain.items.2" = "Album gain";
-/* id="J3e-oT-fy9" */
-"replayGainClip.label" = "Allow the volume gain to clip";
-/* id="HQV-tY-NMj" */
-"replayGainClip.desc" = "If this option is not enabled, clipping will be prevented by lowering the gain.";
-/* key="2yO-To-el7" */
-"$DefaultAuto" = "Default: 0 (Auto)";
-/* key="44m-ge-naq" */
-"maxVolume.label" = "Maximum volume";
-/* key="cJP-ep-847" */
-"maxVolume.desc" = "(100-1000)";
-/* key="50C-1R-wlJ" */
-"videoThreads.label" = "Number of threads";
-/* key="m5P-5f-5uo" */
-"$Video" = "Video";
-/* key="nwg-2K-G2G" */
-"hardwareDecoder.label" = "Hardware decoder";
-/* key="sLR-vV-xn2" */
-"preferredAudioDevice.label" = "Preferred audio device";
-/* key="sV5-dL-MUt" */
-"$Audio" = "Audio";
-/* key="up6-6A-4sn" */
-"$SPDIFOutput" = "S/PDIF output";
-/* key="wjt-O6-KbC" */
-"$PreferredLanguage" = "Preferred language";
-/* key="mNn-cb-IVw" */
-"$EnableToneMappingImagesOnto" = "Enable tone mapping images onto the display";
-/* key="xVk-OA-kOf" */
-"toneMappingTargetPeak.label" = "Target peak";
-/* key="baO-xy-7hD" */
-"$nits" = "nits";
-/* key="aTt-aZ-rjK" */
-"toneMappingTargetPeak.desc" = "Measured peak brightness of the display. If set to 0, IINA detects this value. If detection fails, IINA sets the target peak to 400.";
-/* key="qzI-rq-PqO" */
-"$Algorithm" = "Algorithm";
-/* key="szA-8r-plt" */
-"$AlgorithmToUseForTone" = "Algorithm to use for tone mapping images onto the display.";
-/* key="gfq-nW-dDu" */
-"$ReplayGain" = "Replay Gain";
-/* key="ZRL-b0-k1L" */
-"replayGain.label" = "Gain adjustment";
-/* key="y61-Ah-c92" */
-"replayGain.desc" = "Adjust volume gain according to ReplayGain values stored in the file metadata.";
-/* key="9Q0-hw-PdU" */
-"replayGainPreamp.label" = "Preamp gain";
-/* key="Bew-Pv-igg" */
-"$dB" = "dB";
-/* key="rIW-XP-zGZ" */
-"replayGainPreamp.desc" = "Pre-amplification gain to apply to the ReplayGain gain.";
-/* key="fOz-6g-MhS" */
-"replayGainFallback.label" = "Fallback";
-/* key="m2n-yd-a4n" */
-"replayGainFallback.desc" = "Gain to apply if the file has no ReplayGain tags.";
-/* key="Zco-Ba-Idk" */
-"audioDriverEnableAVFoundation.label" = "Audio driver";
-/* key="fSC-Em-JR4" */
-"audioDriverEnableAVFoundation.items.0" = "Core Audio";
-/* key="" */
-"audioDriverEnableAVFoundation.items.0.desc" = "The default audio driver.";
-/* key="GvK-IC-26c" */
-"audioDriverEnableAVFoundation.items.1" = "AVFoundation";
-/* key="2Md-pv-gSi" */
-"audioDriverEnableAVFoundation.items.1.desc" = "AVFoundation is a new, experimental audio driver offering Spatial Audio support. However, you may encounter some issues when using AVFoundation, such as audio delays.";
-/* key="hwdec.no" */
-"hardwareDecoder.items.0.desc" = "Disable hardware decoding.";
-/* key="hwdec.auto" */
-"hardwareDecoder.items.1.desc" = "Enable hardware decoding. However, most of the video filters won’t work properly.";
-/* key="hwdec.auto-copy" */
-"hardwareDecoder.items.2.desc" = "Enable hardware decoding. It will consume double the memory but work with video filters.";
-
-"gaplessAudio.label" = "Gapless audio";
-"gaplessAudio.desc" = "Gapless audio";
-"gaplessAudio.items.0" = "Disabled";
-"gaplessAudio.items.1" = "Weak";
-"gaplessAudio.items.2" = "Strong";
-"gaplessAudio.items.0.desc" = "Playback of consecutive audio files may result in silence or disruption at the point of file change.";
-"gaplessAudio.items.1.desc" = "If the audio format the decoder output changes, the audio device is closed and reopened, disrupting gapless playback.";
-"gaplessAudio.items.2.desc" = "The audio device is opened using parameters chosen for the first file played and is then kept open for gapless playback. Following files may be resampled.";
-
