+++ iina/SettingsVideoLocalizable.strings
@@ -1,0 +1,48 @@
+/* Video threads */
+"videoThreads.label" = "Number of threads";
+"videoThreads.desc" = "Default: 0 (Auto)";
+
+/* Hardware decoder */
+"hardwareDecoder.label" = "Hardware decoder";
+"hardwareDecoder.items.0" = "Disabled";
+"hardwareDecoder.items.1" = "Auto";
+"hardwareDecoder.items.2" = "Auto (Copy)";
+"hardwareDecoder.items.0.desc" = "Disable hardware decoding.";
+"hardwareDecoder.items.1.desc" = "Enable hardware decoding. However, most of the video filters won't work properly.";
+"hardwareDecoder.items.2.desc" = "Enable hardware decoding. It will consume double the memory but work with video filters.";
+
+/* Dedicated GPU */
+"forceDedicatedGPU.label" = "Force dedicated GPU";
+"forceDedicatedGPU.desc" = "Always use the dedicated GPU for rendering (if it exists). This can improve performance but may reduce battery life.";
+
+/* ICC profile */
+"loadIccProfile.label" = "Load ICC profile";
+"loadIccProfile.desc" = "Load the ICC profile for current display and use it to transform video RGB to screen output. (Only for SDR mode)";
+
+/* HDR */
+"enableHdrSupport.label" = "Enable HDR mode";
+"enableHdrSupport.desc" = "Enable HDR mode by default when playing HDR videos.";
+
+/* Tone mapping */
+"enableToneMapping.label" = "Enable tone mapping";
+"toneMappingTargetPeak.label" = "Target peak";
+"toneMappingTargetPeak.desc" = "Measured peak brightness of the display. If set to 0, IINA detects this value. If detection fails, IINA sets the target peak to 400.";
+"toneMappingAlgorithm.label" = "Algorithm";
+"toneMappingAlgorithm.items.0" = "auto";
+"toneMappingAlgorithm.items.1" = "clip";
+"toneMappingAlgorithm.items.2" = "mobius";
+"toneMappingAlgorithm.items.3" = "reinhard";
+"toneMappingAlgorithm.items.4" = "hable";
+"toneMappingAlgorithm.items.5" = "bt.2390";
+"toneMappingAlgorithm.items.6" = "gamma";
+"toneMappingAlgorithm.items.7" = "linear";
+
+/* Section titles and inline labels */
+"$Video" = "Video";
+"$nits" = "nits";
+"$DefaultAuto" = "Default: 0 (Auto)";
+"$EnableToneMappingImagesOnto" = "Enable tone mapping images onto the display";
+"$Algorithm" = "Algorithm";
+"$AlgorithmToUseForTone" = "Algorithm to use for tone mapping images onto the display.";
+"$Decoding" = "Decoding";
+"$VideoColor" = "Color";
