+++ iina/en.lproj/Localizable.strings
@@ -152,6 +152,8 @@
 "preference.general" = "General";
 "preference.ui" = "Interface";
 "preference.video_audio" = "Video & Audio";
+"preference.video" = "Video";
+"preference.audio" = "Audio";
 "preference.subtitle" = "Subtitle";
 "preference.network" = "Network";
 "preference.control" = "Control";
